# src/Harri.SchoolDemoAPI/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Demo API about students, schools, and students applications to schools",
    "title" : "Swagger School ADMIN - OpenAPI 3.0",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/school/{schoolId}" : {
      "delete" : {
        "description" : "Delete a school by schoolId",
        "operationId" : "deleteSchool",
        "parameters" : [ {
          "description" : "ID of school to delete",
          "explode" : false,
          "in" : "path",
          "name" : "schoolId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/SchoolId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "409" : {
            "description" : "Conflict. School has applications which must be deleted first"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "School not found"
          }
        },
        "summary" : "Delete a school",
        "tags" : [ "school" ]
      },
      "get" : {
        "description" : "Get a school by schoolId",
        "operationId" : "getSchool",
        "parameters" : [ {
          "description" : "ID of school to return",
          "explode" : false,
          "in" : "path",
          "name" : "schoolId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/SchoolId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/School"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "School not found"
          }
        },
        "summary" : "Get a school",
        "tags" : [ "school" ]
      }
    },
    "/school" : {
      "post" : {
        "description" : "Add a new school",
        "operationId" : "addSchool",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewSchool"
              }
            }
          },
          "description" : "Create a new school",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchoolId"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          }
        },
        "summary" : "Add a new school",
        "tags" : [ "school" ]
      },
      "put" : {
        "description" : "Update an existing school by Id",
        "operationId" : "updateSchool",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/School"
              }
            }
          },
          "description" : "Update an existent school",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "School not found"
          }
        },
        "summary" : "Update an existing school",
        "tags" : [ "school" ]
      }
    },
...
    "/application/{applicationId}" : {
      "delete" : {
        "description" : "Delete an application by applicationId",
        "operationId" : "deleteApplication",
        "parameters" : [ {
          "description" : "ID of application to delete",
          "explode" : false,
          "in" : "path",
          "name" : "applicationId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ApplicationId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Application not found"
          }
        },
        "summary" : "Delete an application",
        "tags" : [ "application" ]
      },
      "get" : {
        "description" : "Get an application by applicationId",
        "operationId" : "getApplication",
        "parameters" : [ {
          "description" : "ID of application to return",
          "explode" : false,
          "in" : "path",
          "name" : "applicationId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ApplicationId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Application"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Application not found"
          }
        },
        "summary" : "Get an application",
        "tags" : [ "application" ]
      }
    },
    "/application" : {
      "post" : {
        "description" : "Add a new application",
        "operationId" : "addApplication",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewApplication"
              }
            }
          },
          "description" : "Create a new application",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplicationId"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          }
        },
        "summary" : "Add a new application",
        "tags" : [ "application" ]
      },
      "put" : {
        "description" : "Update an existing application by Id",
        "operationId" : "updateApplication",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Application"
              }
            }
          },
          "description" : "Update an existent application",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Application not found"
          }
        },
        "summary" : "Update an existing application",
        "tags" : [ "application" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "School" : {
        "example" : {
          "schoolId" : 1001,
          "state" : "VIC",
          "schoolName" : "Melbourne Future School",
          "enrollment" : 20000
        },
        "properties" : {
          "schoolId" : {
            "example" : 1001,
            "format" : "int64",
            "type" : "integer"
          },
          "schoolName" : {
            "example" : "Melbourne Future School",
            "type" : "string"
          },
          "state" : {
            "$ref" : "#/components/schemas/State"
          },
          "enrollment" : {
            "example" : 20000,
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "schoolId" ],
        "type" : "object"
      },
      "NewSchool" : {
        "example" : {
          "state" : "VIC",
          "schoolName" : "Melbourne Future School",
          "enrollment" : 20000
        },
        "properties" : {
          "schoolName" : {
            "example" : "Melbourne Future School",
            "type" : "string"
          },
          "state" : {
            "$ref" : "#/components/schemas/State"
          },
          "enrollment" : {
            "example" : 20000,
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Application" : {
        "example" : {
          "major" : "Computer Science",
          "decision" : "Y",
          "schoolId" : 1001,
          "applicationId" : 876581,
          "sId" : 1234
        },
        "properties" : {
          "applicationId" : {
            "example" : 876581,
            "format" : "int64",
            "type" : "integer"
          },
          "sId" : {
            "example" : 1234,
            "format" : "int64",
            "type" : "integer"
          },
          "schoolId" : {
            "example" : 1001,
            "format" : "int64",
            "type" : "integer"
          },
          "major" : {
            "example" : "Computer Science",
            "type" : "string"
          },
          "decision" : {
            "$ref" : "#/components/schemas/Decision"
          }
        },
        "required" : [ "applicationId" ],
        "type" : "object"
      },
      "NewApplication" : {
        "example" : {
          "major" : "Computer Science",
          "decision" : "Y",
          "schoolId" : 1001,
          "sId" : 1234
        },
        "properties" : {
          "sId" : {
            "example" : 1234,
            "format" : "int64",
            "type" : "integer"
          },
          "schoolId" : {
            "example" : 1001,
            "format" : "int64",
            "type" : "integer"
          },
          "major" : {
            "example" : "Computer Science",
            "type" : "string"
          },
          "decision" : {
            "$ref" : "#/components/schemas/Decision"
          }
        },
        "type" : "object"
      },
...
      "SchoolId" : {
        "example" : 1001,
        "format" : "int64",
        "type" : "integer"
      },
      "ApplicationId" : {
        "example" : 876581,
        "format" : "int64",
        "type" : "integer"
      },
      "State" : {
        "description" : "List of valid states",
        "enum" : [ "QLD", "NSW", "VIC", "ACT", "TAS", "WA", "NT", "SA" ],
        "example" : "VIC",
        "type" : "string"
      },
      "Decision" : {
        "description" : "List of valid decisions a school can make on an application",
        "enum" : [ "Y", "N" ],
        "example" : "Y",
        "nullable" : true,
        "type" : "string"
      }
    }
  }
}